Consolidate tests into single file
